backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
alembic==1.17.2
    # via identity-platform-backend (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   starlette
    #   watchfiles
asgiref==3.11.0
    # via opentelemetry-instrumentation-asgi
asyncpg==0.31.0
    # via identity-platform-backend (pyproject.toml)
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via uvicorn
cryptography==46.0.3
    # via identity-platform-backend (pyproject.toml)
fastapi==0.124.2
    # via identity-platform-backend (pyproject.toml)
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via anyio
importlib-metadata==8.7.0
    # via opentelemetry-api
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
opentelemetry-api==1.39.0
    # via
    #   identity-platform-backend (pyproject.toml)
    #   opentelemetry-exporter-prometheus
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-prometheus==0.60b0
    # via identity-platform-backend (pyproject.toml)
opentelemetry-instrumentation==0.60b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-asgi==0.60b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.60b0
    # via identity-platform-backend (pyproject.toml)
opentelemetry-instrumentation-logging==0.60b0
    # via identity-platform-backend (pyproject.toml)
opentelemetry-instrumentation-sqlalchemy==0.60b0
    # via identity-platform-backend (pyproject.toml)
opentelemetry-sdk==1.39.0
    # via
    #   identity-platform-backend (pyproject.toml)
    #   opentelemetry-exporter-prometheus
opentelemetry-semantic-conventions==0.60b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-sdk
opentelemetry-util-http==0.60b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
packaging==25.0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-sqlalchemy
prometheus-client==0.23.1
    # via opentelemetry-exporter-prometheus
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   identity-platform-backend (pyproject.toml)
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via identity-platform-backend (pyproject.toml)
pyjwt==2.10.1
    # via identity-platform-backend (pyproject.toml)
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via identity-platform-backend (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
sqlalchemy==2.0.45
    # via
    #   alembic
    #   identity-platform-backend (pyproject.toml)
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.38.0
    # via identity-platform-backend (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-sqlalchemy
zipp==3.23.0
    # via importlib-metadata
